MRS "Матч", "Статус и История Матча, Метаданные Матча" - https://freetonleague.monday.com/boards/1109042794/pulses/1115672813 1. Изменен тип поля для хранения индикаторов игровой дисциплины. 2. Приложена необходимая миграция таблицы, а также УДАЛЕНИЕ всех настроек игровых дисциплин и ВСЕХ турниров (+ связные сущности)

+++ src/main/java/com/freetonleague/core/domain/enums/GameIndicatorType.java
@@ -9,7 +9,7 @@ public enum GameIndicatorType {
     // tag::common indicators[]
     FRAG_COUNT("frag count", GameIndicatorValueClassType.INTEGER, null),
     FLAG_CAPTURED("flag captured", GameIndicatorValueClassType.BOOLEAN, null),
-    FLAG_CAPTURED_COUNT("flag captured count", GameIndicatorValueClassType.INTEGER, null),
+    FLAG_CAPTURED_COUNT("flag captured count", GameIndicatorValueClassType.DOUBLE, null),
     CHECK_POINT_PASSED("check point passed", GameIndicatorValueClassType.BOOLEAN, null),
     KEY_WORD_DETECTED("found key word", GameIndicatorValueClassType.STRING, null)
     // end::common indicators[]
